Revert "Slack repoter spec"
