+++ spec/factories/users.rb
@@ -1,12 +1,16 @@
 FactoryBot.define do
   factory :user do
-    sequence(:email){|i| "example#{i}@example.com" }
-    password 'pass-word000'
-    sequence(:name){|i| "user#{i}"}
+    sequence(:email){|i| "example#{i}@example.com"}
+    password "pass-word0000"
+    password_confirmation "pass-word0000"
     address { SecureRandom.hex }
     thx_balance 1000
-    sequence(:received_thx){|i| i*100}
-    sequence(:slack_user_id){|i| "#{i}"}
-    sequence(:slack_team_id){|i| 'a' }
+    status "enable"
+  end
+
+  trait :sign_in do
+    after(:create) do |user|
+      create(:access_token, user: user)
+    end
   end
 end
